requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
google-generativeai>=0.8.0
google-genai